refactor(app): refactor css to scsw
