+++ package.json
@@ -51,6 +51,7 @@
     "react": "~17.0.2",
     "react-dom": "~17.0.2",
     "react-markdown": "~5.0.3",
+    "sass": "^1.32.8",
     "serve": "~11.3.2"
   },
   "devDependencies": {
