+++ pages/home.scss
