+++ shared/components/footer/footer.module.scss
