+++ shared/components/paginator/paginator.module.scss
